added check_strong_password function

+++ public/register.php
@@ -1,6 +1,7 @@
 <?php
 
 require('./libs/bcrypt.php');
+require('./libs/functions.php');
 
 $token_entered = $_POST['token'];
 $user_entered = strtolower($_POST['user']);
@@ -10,44 +11,48 @@ if( ($user_entered == '')  || ($password_entered == '') || ($token_entered == ''
     header('Location: index.php');
 }
 else {
+    if(check_strong_password($password_entered)){
+		 
+        require_once($_SERVER['DOCUMENT_ROOT'].'/config/db.php');
 
-    require_once($_SERVER['DOCUMENT_ROOT'].'/config/db.php');
-
-    $token_stmt = $GLOBALS['dbh']->prepare("SELECT token, issued FROM `opsec_registration_tokens` WHERE token = :token");
+        $token_stmt = $GLOBALS['dbh']->prepare("SELECT token, issued FROM `opsec_registration_tokens` WHERE token = :token");
     
-    $token_stmt->execute(array(':token' => $token_entered));
-    $row = $token_stmt->fetch();
-    $token_from_table = $row['token'];
-    $issued_epoch = strtotime($row['issued']);
-    $twelve_hours_ago_epoch = strtotime('-12 hours');
+        $token_stmt->execute(array(':token' => $token_entered));
+        $row = $token_stmt->fetch();
+        $token_from_table = $row['token'];
+        $issued_epoch = strtotime($row['issued']);
+        $twelve_hours_ago_epoch = strtotime('-12 hours');
 
-    if (($token_from_table != '') && ($issued_epoch > $twelve_hours_ago_epoch)){
+        if (($token_from_table != '') && ($issued_epoch > $twelve_hours_ago_epoch)){
      
-        $bcrypt = new bcrypt(12);        
+            $bcrypt = new bcrypt(12);        
 
-        $hashed_pw = $bcrypt->genHash($password_entered);
+            $hashed_pw = $bcrypt->genHash($password_entered);
 
-        if(strlen($password_entered) < 10){
-		die("Password entered too short (less than 10 characters).");
-	}	
-	if (!preg_match('/^[a-z][A-Z]$/', $password)){
-		die("Password needs to contain at least 1 lower/uppercase character, 1 number, and 1 symbol."); 
-	}
-	else{
-	    try {
-                $stmt = $GLOBALS['dbh']->prepare("INSERT INTO `opsec_users`(user, password_hashed) VALUES (:user, :password_hashed)");
-                $stmt->execute(array(':user' => $user_entered, ':password_hashed' => $hashed_pw));
-	    }catch (Exception $e){
-	        die("Error inserting into db");
-            }
-	}
-    }
+            if(strlen($password_entered) < 10){
+    		    die("Password entered too short (less than 10 characters).");
+	    }	
+	    if (!preg_match('/^[a-z][A-Z]$/', $password)){
+		    die("Password needs to contain at least 1 lower/uppercase character, 1 number, and 1 symbol."); 
+	    }
+	    else{
+	        try {
+                    $stmt = $GLOBALS['dbh']->prepare("INSERT INTO `opsec_users`(user, password_hashed) VALUES (:user, :password_hashed)");
+                    $stmt->execute(array(':user' => $user_entered, ':password_hashed' => $hashed_pw));
+	        }catch (Exception $e){
+	            die("Error inserting into db");
+                }
+	    }
+        }
    
-    $delete_token_stmt = $GLOBALS['dbh']->prepare("DELETE FROM `opsec_registration_tokens` WHERE token = :token");
-    $delete_token_stmt->execute(array(':token' => $token_entered));
- 
-    header('Location: index.php');
+        $delete_token_stmt = $GLOBALS['dbh']->prepare("DELETE FROM `opsec_registration_tokens` WHERE token = :token");
+        $delete_token_stmt->execute(array(':token' => $token_entered));
 
+	header('Location: index.php');
+    }
+    else{
+        die();
+    }
 }
 ?>
